Forgot to change some "long" --> long long

+++ mtserver/ClientReceiver.h
@@ -11,7 +11,7 @@ http://stackoverflow.com/questions/39931468/what-does-mean-when-used-in-a-g-argu
 class ClientReceiver : public Thread{
 	private:
 		TCPStream* stream;
-		wqueue<long>& m_queue;
+		wqueue<long long>& m_queue;
 
 	public:
 		ClientReceiver(TCPStream* s, wqueue<long long>& queue);
